kernels can now have a callback that is fired when the kernel's execution is finished
PACXX provides a wrapper for std::promise which hold a bound object that is bound to the lifetime of the promise. when the promise is fulfilled that BindingPromise transfers ownership of the bound object the its future object. This is necessary since CUDA does not allow calls to CUDA API functions from within a Callback (stupid...)